feat: only sync RetroAchievements when new unlocks are detected
Move the RA fetch, additive cache merge, and sync decision into a shared syncRetroAchievements helper used by both triggers: opening a launchbox game page and ending an emulator session. The backend sync is now enqueued only when the local RA fetch finds unlocks missing from the cache, and only for active subscribers. Sync errors are logged instead of swallowed (expected subscription/not-logged-in skips are ignored).
